requirements.txt
keyboard
mailtrap
python-dotenv